backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
djoser>=2.1.0
psycopg2-binary>=2.8
djangorestframework-simplejwt>=5.2.2
